requirements.txt
py-cord==2.6.1
aiohttp
python-dotenv
flask